Add AppendCSharpString extension method for StringBuilder and ToCSharpString extension method for String.
